pyproject.toml
[build-system]
requires = [
    "setuptools>=65", 
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "native-chess-transformer"
version = "0.1.0"
description = "Hybrid CNN+Transformer AlphaZero chess engine"
authors = [
    {name = "Native Chess Transformer Team", email = "nct@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"

# Core runtime dependencies
dependencies = [
    "torch>=2.2",
    "tqdm>=4.66",
    "numpy>=1.23",
    "python-chess>=1.999",
    "webdataset>=0.2.63",
    "PyYAML>=6.0",
    "httpx>=0.27"
]

[project.optional-dependencies]
# Extras for development, visualization, etc.
wandb = ["wandb>=0.16"]
dev = [
    "pytest>=7.4",
    "ruff>=0.4",
    "black>=24.4",
    "tensorboard>=2.13"
]
web = [
    "Flask>=2.3",
    "flask-cors>=4.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests*", "examples*", "docs*"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

